@@ -2,12 +2,10 @@
   "name": "nemis_api",
   "version": "0.0.1-alpha",
   "description": "An un-official NEMIS api aimed toward bulk operations and time saving",
-  "main": "index.js",
+  "main": "./dist/index.js",
   "type": "commonjs",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/index.js"
-  },
+  "author": "d1dee",
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "github.com/d1dee/nemisApi"
@@ -17,8 +15,29 @@
     "KENYA",
     "EDUCATION"
   ],
-  "author": "d1dee",
-  "license": "MIT",
+  "scripts": {
+    "build": "tsc",
+    "dev:tsc": "tsc -w | nodemon --inspect --watch ./dist/ ",
+    "start": "node dist/index.js"
+  },
+  "_moduleAliases": {
+    "@database": 
+      "./dist/src/database"
+    ,
+    "@routes": 
+      "./dist/src/routes"
+    ,
+    "@middleware": 
+      "./dist/src/middleware"
+    ,
+    "@controller": 
+      "./dist/src/controller"
+    ,
+    "@libs": 
+      "./dist/src/libs",
+      "@types":"./types"
+    
+  },
   "dependencies": {
     "axios": "^1.3.3",
     "crypto-js": "^4.1.1",
@@ -27,6 +46,7 @@
     "express-fileupload": "^1.4.0",
     "form-data": "^4.0.0",
     "jsonwebtoken": "^9.0.0",
+    "module-alias": "^2.2.3",
     "moment": "^2.29.4",
     "mongoose": "^6.6.5",
     "node-html-parser": "^6.1.4",
@@ -42,7 +62,8 @@
     "@types/express-fileupload": "^1.4.1",
     "@types/jsonwebtoken": "^9.0.1",
     "@types/node": "^18.7.13",
+    "nodemon": "^3.0.1",
     "prettier": "^2.8.1",
-    "typescript": "^5.0.4"
+    "typescript": "^5.2.2"
   }
 }
