@@ -6,7 +6,7 @@ import { Request } from 'express';
 import learner from '../../database/learner';
 import { Grades, ListLearner, SearchLearnerApiResponses } from '../../../types/nemisApiTypes';
 import CustomError from '../../libs/error_handler';
-import { sendErrorMessage } from '../utils/middlewareErrorHandler';
+import { sendErrorMessage } from '../utils/middleware_error_handler';
 import { NemisWebService } from '../../libs/nemis/nemis_web_handler';
 import { NemisApiService } from '../../libs/nemis/nemis_api_handler';
 import { uniqueIdentifierSchema } from '../../libs/zod_validation';
@@ -101,7 +101,7 @@ const captureContinuingLearner = async (req: Request) => {
 			learnerToCapture.map(learner =>
 				learner.birthCertificateNo
 					? nemisApi.searchLearner(encodeURI(learner.birthCertificateNo))
-					: Promise.reject(new CustomError('Learner has no valid birthCertificate'))
+					: Promise.reject(new CustomError('Learner has no valid birthCertificate', 500))
 			)
 		);
 
