@@ -2,12 +2,12 @@
  * Copyright (c) 2023. MIT License.  Maina Derrick
  */
 
-import { Request, Response } from "express";
-import logger from "../../libs/logger";
+import { Request, Response } from 'express';
+import logger from '../../libs/logger';
 
 interface SetHeaders {
     Authorization: string;
-    "Access-Control-Expose-Headers": string;
+    'Access-Control-Expose-Headers': string;
     Expires: string;
 }
 
@@ -35,74 +35,76 @@ export default class {
     //Generate generic error response
     error(errorCode: number, message?: string, cause?: any) {
         //check if a response has been sent
+        if (process.env.NODE_ENV === 'production') cause = undefined;
+
         switch (errorCode) {
             case 400:
                 this.response.status(400).json({
                     success: false,
-                    message: message || "Bad Request",
+                    message: message || 'Bad Request',
                     cause: cause || undefined
                 });
                 break;
             case 401:
                 this.response.status(401).json({
                     success: false,
-                    message: message || "Unauthorized",
+                    message: message || 'Unauthorized',
                     cause: cause || undefined
                 });
                 break;
             case 403:
                 this.response.status(403).json({
                     success: false,
-                    message: message || "Forbidden",
+                    message: message || 'Forbidden',
                     cause: cause || undefined
                 });
                 break;
             case 404:
                 this.response.status(404).json({
                     success: false,
-                    message: message || "Not Found",
+                    message: message || 'Not Found',
                     cause: cause || undefined
                 });
                 break;
             case 405:
                 this.response.status(405).json({
                     success: false,
-                    message: message || "Method Not Allowed",
+                    message: message || 'Method Not Allowed',
                     cause: cause || undefined
                 });
                 break;
             case 408:
                 this.response.status(408).json({
                     success: false,
-                    message: message || "Request Timeout",
+                    message: message || 'Request Timeout',
                     cause: cause || undefined
                 });
                 break;
             case 409:
                 this.response.status(409).json({
                     success: false,
-                    message: message || "Conflict",
+                    message: message || 'Conflict',
                     cause: cause || undefined
                 });
                 break;
             case 413:
                 this.response.status(413).json({
                     success: false,
-                    message: message || "Payload Too Large",
+                    message: message || 'Payload Too Large',
                     cause: cause || undefined
                 });
                 break;
             case 415:
                 this.response.status(415).json({
                     success: false,
-                    message: message || "Unsupported Media Type",
+                    message: message || 'Unsupported Media Type',
                     cause: cause || undefined
                 });
                 break;
             case 429:
                 this.response.status(429).json({
                     success: false,
-                    message: message || "Too Many Requests",
+                    message: message || 'Too Many Requests',
                     cause: cause || undefined
                 });
                 break;
@@ -110,42 +112,42 @@ export default class {
             case 500:
                 this.response.status(500).json({
                     success: false,
-                    message: message || "Internal Server Error",
+                    message: message || 'Internal Server Error',
                     cause: cause || undefined
                 });
                 break;
             case 502:
                 this.response.status(502).json({
                     success: false,
-                    message: message || "Bad Gateway",
+                    message: message || 'Bad Gateway',
                     cause: cause || undefined
                 });
                 break;
             case 503:
                 this.response.status(503).json({
                     success: false,
-                    message: message || "Service Unavailable",
+                    message: message || 'Service Unavailable',
                     cause: cause || undefined
                 });
                 break;
             case 504:
                 this.response.status(504).json({
                     success: false,
-                    message: message || "Gateway Timeout",
+                    message: message || 'Gateway Timeout',
                     cause: cause || undefined
                 });
                 break;
             case 505:
                 this.response.status(505).json({
                     success: false,
-                    message: message || "HTTP Version Not Supported",
+                    message: message || 'HTTP Version Not Supported',
                     cause: cause || undefined
                 });
                 break;
             case 511:
                 this.response.status(511).json({
                     success: false,
-                    message: message || "Network Authentication Required",
+                    message: message || 'Network Authentication Required',
                     cause: cause || undefined
                 });
                 break;
@@ -163,23 +165,26 @@ export default class {
     //method used to send a response to the client
     respond(data: any, message?: string, statusCode?: number) {
         statusCode = statusCode || 200;
+
+        // todo: Walk through data object and remove sensitive data
+        
         //check if a response has been sent
         if (this.response?.headersSent) {
-            return logger.warn("Headers sent");
+            return logger.warn('Headers sent');
         }
         //check if response is an error
         if (data instanceof Error) {
             //logger.error(data);
             this.response.status(500).send({
                 success: false,
-                message: "Internal Server Error",
+                message: 'Internal Server Error',
                 cause: data.cause || undefined
             });
         } else {
             //send the response to the client
             this.response.status(statusCode).send({
                 success: true,
-                message: message || "Operation complete successfully",
+                message: message || 'Operation complete successfully',
                 data: data || []
             });
         }
