@@ -8,7 +8,7 @@ import { deleteInstitution, getInstitution, updateInstitution } from '../../midd
 const institution = Router().use(json());
 
 institution.get('/', getInstitution);
-institution.patch('/update', updateInstitution);
+institution.post('/update', updateInstitution);
 institution.delete('/delete', deleteInstitution);
 
 export default institution;
