@@ -6,7 +6,7 @@ import { JwtPayload } from 'jsonwebtoken';
 import send_response from '../../src/middleware/utils/send_response';
 import { DatabaseInstitution, DatabaseToken } from '../nemisApiTypes';
 import { NemisWebService } from '../../src/libs/nemis/nemis_web_handler';
-import { z as zod } from 'zod/lib';
+import { z as zod } from 'zod';
 import { queryParameterSchema } from '../../src/middleware/utils/query_params';
 
 declare module 'express-serve-static-core' {
